Conflicts:
	Common/acgui.h
	Common/acplatfm.h
	Common/acroom.h
	Common/acruntim.h
	Common/compress.h
	Common/cscomp.h
	Common/csrun.cpp
	Common/util/compress.cpp
	Engine/Makefile.linux
	Engine/ac.cpp
	Engine/acchars.cpp
	Engine/acfonts.cpp
	Engine/acplatfm.cpp
	Engine/platform/linux/acpllnx.cpp
	Engine/plugin/agsplugin.h
	debian/rules

+++ Common/ac/interaction.cpp
@@ -84,8 +84,8 @@ void NewInteractionCommand::ReadFromFile(FILE *fp)
         data[i].ReadFromFile(fp);
     }
     // all that matters is whether or not these are null...
-    children = (NewInteractionAction *) getw(fp);
-    parent = (NewInteractionCommandList *) getw(fp);
+    children = (NewInteractionAction *)(long)getw(fp);
+    parent = (NewInteractionCommandList *)(long)getw(fp);
 //#else
 //    throw "NewInteractionCommand::ReadFromFile() is not implemented for little-endian platforms and should not be called.";
 //#endif
@@ -100,8 +100,8 @@ void NewInteractionCommand::WriteToFile(FILE *fp)
     {
         data[i].WriteToFile(fp);
     }
-    putw((int)children, fp);
-    putw((int)parent, fp);
+    putw((long)children, fp);
+    putw((long)parent, fp);
 //#else
 //    throw "NewInteractionCommand::WriteToFile() is not implemented for little-endian platforms and should not be called.";
 //#endif
@@ -207,8 +207,10 @@ void serialize_new_interaction (NewInteraction *nint, FILE*ooo) {
   putw (1, ooo);  // Version
   putw (nint->numEvents, ooo);
   fwrite (&nint->eventTypes[0], sizeof(int), nint->numEvents, ooo);
+
+  // 64 bit: The pointer is only checked against NULL to determine whether the event exists
   for (a = 0; a < nint->numEvents; a++)
-    putw ((int)nint->response[a], ooo);
+    putw ((long)nint->response[a], ooo);
 
   for (a = 0; a < nint->numEvents; a++) {
     if (nint->response[a] != NULL)
@@ -251,6 +253,8 @@ NewInteraction *deserialize_new_interaction (FILE *ooo) {
   }
   fread (&nitemp->eventTypes[0], sizeof(int), nitemp->numEvents, ooo);
   //fread (&nitemp->response[0], sizeof(void*), nitemp->numEvents, ooo);
+
+  // 64 bit: The pointer is only checked against NULL to determine whether the event exists
   for (a = 0; a < nitemp->numEvents; a++)
     nitemp->response[a] = (NewInteractionCommandList*)getw(ooo);
