+++ Engine/ac/dynamicsprite.cpp
@@ -161,7 +161,7 @@ void DynamicSprite_CopyTransparencyMask(ScriptDynamicSprite *sds, int sourceSpri
     int bytesPerPixel = (colDep + 1) / 8;
 
     unsigned short *shortPtr;
-    unsigned long *longPtr;
+    unsigned int *longPtr;
     for (int y = 0; y < target->h; y++)
     {
         unsigned char * sourcePixel = source->line[y];
@@ -169,7 +169,7 @@ void DynamicSprite_CopyTransparencyMask(ScriptDynamicSprite *sds, int sourceSpri
         for (int x = 0; x < target->w; x++)
         {
             shortPtr = (unsigned short*)sourcePixel;
-            longPtr = (unsigned long*)sourcePixel;
+            longPtr = (unsigned int*)sourcePixel;
 
             if ((colDep == 8) && (sourcePixel[0] == maskColor))
             {
@@ -185,7 +185,7 @@ void DynamicSprite_CopyTransparencyMask(ScriptDynamicSprite *sds, int sourceSpri
             }
             else if ((bytesPerPixel == 4) && (longPtr[0] == maskColor))
             {
-                ((unsigned long*)targetPixel)[0] = maskColor;
+                ((unsigned int*)targetPixel)[0] = maskColor;
             }
             else if ((bytesPerPixel == 4) && (sourceHasAlpha))
             {
