+++ Engine/ac/timer.cpp
@@ -4,7 +4,7 @@
 
 extern volatile int mvolcounter;
 
-unsigned long loopcounter=0,lastcounter=0;
+unsigned int loopcounter=0,lastcounter=0;
 volatile unsigned long globalTimerCounter = 0;
 
 volatile int timerloop=0;
